--- src/components/CourseDescription/style.module.css
@@ -1,11 +1,0 @@
-.main {
-  width: 1440px;
-  padding-left: 140px;
-  padding-right: 140px;
-  padding-bottom: 95px;
-}
-
-.title {
-  margin-top: 75px;
-  margin-bottom: 40px;
-}
