+++ src/components/LoginAvatar/style.module.css
@@ -3,3 +3,10 @@
   gap: 13px;
   align-items: center;
 }
+
+.link {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  text-decoration: none;
+}
