+++ src/index.css
@@ -2,11 +2,10 @@
   margin: 0;
   padding: 0;
   box-sizing: border-box;
+  font-family: 'SansSerif', sans-serif;
 }
 
 body {
-  font-family: -apple-system, 'StratosSkyeng', 'Abhaya Libre ExtraBold', 'Segoe UI', 'Roboto',
-    'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
