--- src/layouts/style.module.css
@@ -1,6 +1,0 @@
-.container {
-  width: 100%;
-  padding-left: calc(50% - 580px);
-  padding-right: calc(50% - 580px);
-  background: #271a58;
-}
