--- src/pages/CourseDescriptionPages/Header/Header.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import styles from './style.module.css';
-import Logo from '../../../components/Logo/Logo';
-
-export const Header = () => {
-  return (
-    <div className={styles.logo}>
-      <Logo color="#000" />
-    </div>
-  );
-};
