--- src/pages/CourseDescriptionPages/Header/style.module.css
@@ -1,4 +1,0 @@
-.logo {
-  margin-top: 30px;
-  margin-bottom: 75px;
-}
