+++ src/pages/CourseDescriptionPages/style.module.css
@@ -1,0 +1,12 @@
+.main {
+  min-height: 100vh;
+  width: 100%;
+  padding-left: calc(50% - 580px);
+  padding-right: calc(50% - 580px);
+  background: #ffffff;
+}
+
+.title {
+  margin-top: 75px;
+  margin-bottom: 40px;
+}
