--- src/pages/CourseProgressPage/CourseHeader/CourseHeader.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import style from './style.module.css';
-
-const CourseHeader = () => {
-  return (
-    <header className={style.header}>
-      <div>Logo</div>
-      <div>User</div>
-    </header>
-  );
-};
-
-export default CourseHeader;
