--- src/pages/CourseProgressPage/CourseHeader/style.module.css
@@ -1,8 +1,0 @@
-.header {
-  width: 100%;
-  margin-top: 24px;
-  margin-bottom: 66px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
