+++ src/pages/CourseProgressPage/style.module.css
@@ -1,8 +1,12 @@
 .wrapper {
+  min-height: 100vh;
+  width: 100%;
   display: flex;
   flex-direction: column;
   align-items: center;
-  background-color: #ffffff;
+  padding-left: calc(50% - 580px);
+  padding-right: calc(50% - 580px);
+  background: #ffffff;
 }
 
 .title {
