--- src/pages/MainSkyFitnessProPage/Header/style.module.css
@@ -1,32 +1,0 @@
-.header {
-  padding-top: 30px;
-  padding-bottom: 30px;
-  display: flex;
-  justify-content: space-between;
-}
-
-.button {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 8px;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 24px;
-  width: 77px;
-  height: 36px;
-  color: #fff;
-
-  background: #140d40;
-  border-radius: 46px;
-  border: none;
-  cursor: pointer;
-}
-
-.button:hover {
-  background: #2d1f79;
-}
-
-.button:active {
-  background: #3b29a2;
-}
