+++ src/pages/UserProfile/MyCoursesBlock/MyCoursesContent/style.module.css
@@ -7,3 +7,7 @@
 .cardBlock {
   position: relative;
 }
+
+.link {
+  text-decoration: none;
+}
