--- src/pages/UserProfile/UpperBlock/UpperBlock.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import Logo from '../../../components/Logo/Logo';
-import LoginAvatarBlock from '../../../components/LoginAvatar/LoginAvatarBlock';
-import styles from './styles.module.css';
-
-const UpperBlock = () => {
-  return (
-    <div className={styles.upperBlock}>
-      <Logo color="#000000" />
-      <LoginAvatarBlock />
-    </div>
-  );
-};
-
-export default UpperBlock;
