--- src/pages/UserProfile/UpperBlock/styles.module.css
@@ -1,7 +1,0 @@
-.upperBlock {
-  display: flex;
-  padding-top: 33px;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 74px;
-}
