+++ src/pages/UserProfile/UserProfile.jsx
@@ -1,14 +1,14 @@
 import React from 'react';
-import UpperBlock from './UpperBlock/UpperBlock';
 import CentralBlock from './CentralBlock/CentralBlock';
 import MyCoursesBlock from './MyCoursesBlock/MyCoursesBlock';
 import styles from './style.module.css';
+import Header from '../../components/Header/Header';
 
 const UserProfile = () => {
   return (
     <>
       <div className={styles.userProfile}>
-        <UpperBlock />
+        <Header logoColor="#000000" />
         <CentralBlock />
         <MyCoursesBlock />
       </div>
