+++ src/pages/UserProfile/style.module.css
@@ -1,5 +1,7 @@
-/* удалил отступы , но пусть пока файл останется */
 .userProfile{
-    height: 100vh;
-    width: 100vw;
+    min-height: 100vh;
+    width: 100%;
+    padding-left: calc(50% - 580px);
+    padding-right: calc(50% - 580px);
+    background: #ffffff;
 }
